Move SortKey/SortDirection to app.model.ts; extract displayedMessages logic to app.helpers.ts
Agent-Logs-Url: https://github.com/dev-chat/mocker/sessions/b2c25933-76dd-47f7-a807-9aa04e6eed9f
